BlueRootAi/requirements.txt
Flask
gunicorn
gpiozero
smbus2
PyYAML
python-dotenv